requirements.txt
# Framework principal
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Métricas Prometheus
prometheus-fastapi-instrumentator==7.1.0

# Tracing OpenTelemetry
opentelemetry-distro==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-exporter-otlp-proto-http==1.24.0

# Para testes
pytest==8.2.1
requests==2.31.0